+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3learners.mboost 0.2.1.9000
+
+- Internal changes only.
+
+
 # mlr3learners.mboost 0.2.0.9000
 
 - Use `$help()` method from super class